LayoutTests/http/wpt/loading/early-hints/preconnect-csp-allowed.h2.window-expected.txt

CONSOLE MESSAGE: Preconnecting to https://not-web-platform.test:9000/ due to early hint


PASS Early hints preconnect
